Start working on robot management again
