+++ src/ROS/rosapi.tsx
@@ -1,0 +1,4 @@
+export interface Topics {
+  topics: string[];
+  types: string[];
+}
